moved the .h and .cpp files from the Pathfinding and src sub-folders into the root directory. After, went into the BUILD file to change some code to find the new locations of the files finally, deleted the Pathfinding and src subfolders or subdirectories from extlib/enlsvg
