+++ src/extlibs/enlsvg/BUILD
@@ -2,8 +2,8 @@ package(default_visibility = ["//visibility:public"])
 
 cc_library(
     name = "enlsvg",
-    srcs = glob(["src/*.cpp"]),
-    hdrs = glob(["Pathfinding/*.h"]),
+    srcs = glob(["*.cpp"]),
+    hdrs = glob(["*.h"]),
     linkstatic = True,
     deps = [],
 )
