+++ src/extlibs/enlsvg/ENLSVG.h
@@ -1,0 +1,8 @@
+#ifndef _ENLSVG_H_
+#define _ENLSVG_H_
+
+#include "extlibs/enlsvg/ENLSVGAlgorithm.h"
+#include "extlibs/enlsvg/Grid.h"
+#include "extlibs/enlsvg/PathfindingDataTypes.h"
+
+#endif
