+++ src/extlibs/enlsvg/ENLSVGGraph.cpp
@@ -1,9 +1,9 @@
-#include "extlibs/enlsvg/Pathfinding/ENLSVGGraph.h"
+#include "extlibs/enlsvg/ENLSVGGraph.h"
 
 #include <algorithm>
 
-#include "extlibs/enlsvg/Pathfinding/Grid.h"
-#include "extlibs/enlsvg/Pathfinding/LineOfSightScanner.h"
+#include "extlibs/enlsvg/Grid.h"
+#include "extlibs/enlsvg/LineOfSightScanner.h"
 
 namespace Pathfinding
 {
