+++ src/extlibs/enlsvg/Fraction.cpp
@@ -1,4 +1,4 @@
-#include "extlibs/enlsvg/Pathfinding/Fraction.h"
+#include "extlibs/enlsvg/Fraction.h"
 
 #include <iostream>
 #include <sstream>
