--- src/extlibs/enlsvg/Pathfinding/ENLSVG.h
@@ -1,8 +1,0 @@
-#ifndef _ENLSVG_H_
-#define _ENLSVG_H_
-
-#include "ENLSVGAlgorithm.h"
-#include "Grid.h"
-#include "PathfindingDataTypes.h"
-
-#endif
