+++ src/extlibs/enlsvg/PathfindingDataTypes.h
